ref(feedback/issues): refactor groupActivity.tsx
